test5 parallel & gatconv op

+++ gen_graph.py
@@ -13,17 +13,17 @@ df = pd.read_csv('DATA/{}/edges.csv'.format(args.data))
 num_nodes = max(int(df['src'].max()), int(df['dst'].max())) + 1
 print('num_nodes: ', num_nodes)
 
-int_train_indptr = np.zeros(num_nodes + 1, dtype=np.int)
+int_train_indptr = np.zeros(num_nodes + 1, dtype=np.int32)
 int_train_indices = [[] for _ in range(num_nodes)]
 int_train_ts = [[] for _ in range(num_nodes)]
 int_train_eid = [[] for _ in range(num_nodes)]
 
-int_full_indptr = np.zeros(num_nodes + 1, dtype=np.int)
+int_full_indptr = np.zeros(num_nodes + 1, dtype=np.int32)
 int_full_indices = [[] for _ in range(num_nodes)]
 int_full_ts = [[] for _ in range(num_nodes)]
 int_full_eid = [[] for _ in range(num_nodes)]
 
-ext_full_indptr = np.zeros(num_nodes + 1, dtype=np.int)
+ext_full_indptr = np.zeros(num_nodes + 1, dtype=np.int32)
 ext_full_indices = [[] for _ in range(num_nodes)]
 ext_full_ts = [[] for _ in range(num_nodes)]
 ext_full_eid = [[] for _ in range(num_nodes)]
